requirements.txt
flask==3.0.2
python-dotenv==1.0.1
firebase-admin==6.4.0
requests==2.31.0
beautifulsoup4==4.12.3
flask-login==0.6.3 